Split creator share logic
